atualizando versao

+++ WebMvcSgq/Models/Interface/IProcessoRepositorio.cs
@@ -9,10 +9,10 @@ namespace WebMvcSgq.Models.Interface
     interface IProcessoRepositorio
     {
         IEnumerable<tbl_Processo> GetProcessos();
-        tbl_Processo GetProcessoPorID(int processoId);
+        tbl_Processo GetProcessoPorID(long processoId);
         void AdicionaProcesso(tbl_Processo processo);
         void DeletaProcesso(long processoId);
         void AtualizaProcesso(tbl_Processo processo);
-        tbl_Processo Detalhes(int processoId);
+        tbl_Processo Detalhes(long processoId);
     }
 }
